Add best search algorithm
I'm likely not following function overload convention here but the idea is to have an `MP3Stat.parse` function that doesn't calculate the duration. Duration would be useful if we were utilizing tracks in some way but we really only care about the metadata.
The hope with this change was to pull in Benchee to benchmark each approach but I don't know if I really have the time to do that and create a polished UI. There are a lot of ideas flowing that may mostly be version 3.0 when I need a very minimal MVP (I know the word minimal is used twice there).