Specification v0.1 complete
I have an understanding of how this will look in my head now. I feel like I really had a good pattern going and then I had to charge my phone before I could even capture it in text. I believe this is as close of a representation to that though. Demo day is going to be largely sleight of hand, just call me Jeremy Copperfield.