Add Tailwind config from personal site
I want to use similar branding so I'm pulling the palette from my personal website. Fonts work but there's a flash of unstyled content and setting boldness seems to get replaced. I don't have time to debug why and we can go with the default font loading or try to match it later.