#1 PubSub technically not necessary for this
Propagating from the parent LiveView to a component is as easy as using `send_update`. My problem now comes from wanting to trigger notifications from a background process. The only LiveViews in these instances are the page. The sidebar is a stateless component.

We could use live_session and on_mount to always mount a pubsub subscribe but what catches the events? Each damn page? I need to make Sidebar a LiveView and look into `:sticky` usage. This really fucking sucks honestly. I took enough of a detour just trying to nest the sidebar as a LiveComponent. This helped me greatly: https://stackoverflow.com/a/66325108