Propagate `current_user`
We may need to place even the PageController in the same section. The music page is broken as hell on its own and needs the redirect. This should be no different than home then actually, using its own separate layout. The music event can just be a landing page. I just don't have the energy to style one just yet.