Documentation cleanup
Add index pages and backlinks to tie each file together. I need to really describe what the application should be doing according to what is actually here, not what I came up with at the beginning. I don't believe much has deviated but it's worth repeating it to be sure.