#4 #8 Verify creates an album if missing and notifies \o/
This is messy as shit because I'm down to the wire. At the notification level we're using `~D[]` types already so no need to format, just work with what we got.

What we have should be technically functional. I haven't tested Oban but it works. What doesn't work is updating the `verified_at` which would push this job to run indefinitely. We can just run things in `iex`.