#14 GitHub CI complete
Following https://github.com/felt/ultimate-elixir-ci.
Changed the matrix Elixir/OTP versions to what I use with asdf. There may be a better way to unify this.
Changed postgis database image to stock postgres:14 as that's what Fly.io uses.
I have setup a Gigalixir account and stub for a `beatseek` app since I had so many issues. It may be useful to have this as my PR preview/staging server. Tailwind deploys to Vercel on every commit and it was really slick to see every change without having to do the work locally.

Changing the formatter to 120 characters adjusted just about everything but this is fine. We can always revert back to 80 characters or be explicit.
It's worth noting that they use a `priv/repo/data_migrations` path for explicit data migrations. I'd love to get in the habit of doing them, even for toy projects.
Tests are explicitly commented out for all runs, which kind of negates the nightly check. This is fine though. We'll make an explicit issue to cover making all tests pass (and adding new ones).
Fixed a few warnings I had a hard time seeing with unused MP3Stat pictures and Artist.
Fixed some credo refactor suggestions like removing the redundant `result` checks for the Oban worker. We can just use the case statement or worst-case pipe to a variable and use that.

A lot of modules are missing @moduledoc declarations and doctests. Writing those out will help me with tests.