#20 Year parsing
Taking a play from https://thepugautomatic.com/2016/01/pattern-matching-complex-strings/ to build a date pattern to match on. This catches the YYYY-MM-DD and YYYY-MM-DDTHH:MM:SS cases. Even though I'm US based, we really don't care if the day or month is accurate as we discard the captures to focus on just the year.

Fixed genre to return the genre if the mapping is unsuccessful.