#20 Release date to year
Using the same parsing behavior as the other transformer. There is likely a better way of merging the two transformers and handling the different attribute shapes. I think it's important to note it's a smell that I should be able to refactor later when tests are updated.
Pulling the `spotify_artist` params down so we can get the genre(s) and look to inject the artist's image.