#20 Artist image url
Now I'm just copy/pasting shit everywhere at this point. It is worth pointing out that the dimensions for an artist differ slightly so it really isn't 1:1.
Display an ugly image instead of the icon. As this moves to a component the image will need to have a circular mask applied. I know I'm technically introducing a bug here but it's fine.