#20 Fix album art shrinkage
There's likely a better way here but this is good enough. We don't want to always enforce this and it may be different for grids. Setting this on the `Albums :: Index` page exhibits problematic behavior when resizing the window due to trying to keep a fixed width. Alternatives would be to prevent flex shrinkage where we don't want it. I think explicit dimensions are a good approach for other screens where responsive doesn't matter all that much or at least shouldn't adjust the image really.