#34 Badge jank
Mix format touches
Refactored artist, album, and notification badges as a stateless component. The Sidebar can keep track of all the counts.

LiveComponent is so damn weird here because we want to trigger an event from the parent but it always has to be done through an update where you're sending assigns down. There seems to be either very rigid use cases I'm not capturing or I'm seeing a very weird problem that likely wouldn't happen elsewhere. I wanted a LiveComponent so it would be the source of truth for the counts. My intention was to group all of the similar behavior together but the second I threw in PubSub, everything fell apart.