#31 Scanning directory defaults
Moved Spotify configuration to runtime. I failed to realize the prod environment check for some reason.
Default to `""` if the directory is not supplied or the config variable isn't found. This will result in a blank list `[]` instead of raising an error.
Removed the more verbose `full_scan` function. There's probably a better way of doing this per the comments but this is fine for now.