#19 Fix gen auth tests
Fix awkward login and registration pages by removing the background color. I had stopped caring about this until these tests. This isn't really ready for users but I want to lay the groundwork.
LiveViews needed more temporary assigns
Logging in redirects as the `music` controller action is a redirect. We'll change this later when we have a landing page and logged in landing page.
Fixed `Log in` to `Sign in`. Did I change that? Were these generated with broken tests? The login page doesn't have a register link, we want to discourage registrations even though the routes work.