#19 Transformer tests
Add transformer tests with happy paths only. I've yet to see any unhappy paths hit in practice but it's worth trying to transform blank structs to understand what it could potentially look like.
Add type information to transformers to make it clear that we expect structs.

I luckily found the `struct!()` function when looking in LiveBeats to see how to create fixtures that populated structs like this. It felt possible, but the piping syntax to a struct without a `new` factory function wasn't working. I have a feeling I'm going to be using this a ton as I prefer structs as my goto data type or a map if I don't care to have a formal type.

What does concern me is the use of different map syntax. We need the arrow syntax (`=>`) to use the string keys Spotify returns. The entire structure is all string keys, yet we seemingly have to use this syntax for id and other keys. I had more fields that I had to remove due to a mismatch. The scaffold test fixtures seem to exclusively use `id: "value"`. There may be ways of changing the code so that `Map.get()` always returns the right value I'll try to explore later. I really don't want something this brittle or I need to upgrade my understanding that this is okay.