#19 Scanner tests
Had the bright idea of using `kid3-cli` to inject the metadata from valid mp3s into empty husks that only contain both the v1 and v2 tags. I promise there is no copyrighted material in the files themselves beyond the metadata bytes. I believe according to court cases that this does not mean I'm distributing copyrighted material even though it looks like it.

Our tests don't seem to be testing a lot on the surface though as we're really only verifying the injected structs match the database values. The expected list can always be assumed to be filled in but the actual values could be nil. We're really only calling the setup and assert when we know both sides are filled out as expected.

These tests seem brittle and circular because they are in a way. It's probably better to have multiple fixture outcomes instead of the same directory with the same 3 albums.