#56 Verify single artist
This is a huge smell and I noticed the missing docs. ElixirLS isn't even giving me style suggestions so I probably need to kick VSCode a minute to get them back. There is a case in the return from `Beatseek.Notifications.Delivery` that I'm not taking into account somehow. Though now I wonder if we'll run into that since we're no longer relying on Oban's retry behavior. It may be useful to create another worker to cover each case but this is fine for now. I'm okay with leaky abstractions as long as I can identify and correct them later.