Public release
Add release retrospective and known issue callout since I'd consider it a game-breaking bug.
I think it's in a good place to showcase a project I can see through to this state. Not every project will have the right mix out of the gate but I'll be looking back to this for a lot of what I do in the future.
I wish this would've happened sooner but it's better late than never.