Fix documentation links
Added duped document index since I think it helps when navigating in GitHub.
Light journal cleanup after proofreading again. Added * sections to show changes though that's kind of awkward. Removed my altar idea as it made sense to me but that was a weird ass sentence. To make it simpler, we need money and we see money show up that is unexpected.