#60 Responsive improvements
Remove the functional sidebar component. We can always add it back and the deviation is so great it should probably be rebuilt if it's used like that.
The album art's maximum width is 300px. Adding this to the art itself and the relative text. We can't use the container because of the grid layout.
Remove the sticky and width to the sidebar insides as the container covers everything now.
Add background to footer so it's no longer transparent and difficult to see.
Fix sidebar container by defaulting to full width and clamping the maximum width to xs on small form factors. I had incorrectly set the width for all screens when mobile needed different behavior to be more usable.